don't delete by default

+++ scripts/record_rosbag_filtered.bash
@@ -10,4 +10,8 @@ filtered_bag_name=$current_date_time"_"$reason.bag
 # Filter bag
 rosbag filter $current_date_time.bag $filtered_bag_name "topic not in ['/ugr/car/sensors/cam0/image']"
 
-rm $current_date_time.bag
+read -p "Do you want to delete the original rosbag? (y/n): " delete_original
+
+if [ $delete_original == "y" ]; then
+    rm $current_date_time.bag
+fi
